requirements.txt
affine==2.4.0
appJar==0.94.0
apturl==0.5.2
asn1crypto==1.4.0
asttokens==2.4.1
attrs==23.1.0
bashlex==0.18
bcrypt==3.2.0
black==24.1.1
blinker==1.4
Brlapi==0.8.3
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
compiledb==0.10.1
configobj==5.0.6
confluent-kafka==2.3.0
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fasteners==0.14.1
findspark==2.0.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
fonttools==4.44.3
future==0.18.2
greenlet==2.0.2
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kafka-python==2.0.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
packaging==23.2
panda==0.3.1
pandas==2.0.2
paramiko==2.9.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pycairo==1.20.1
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyspark==3.5.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
rasterio==1.3.9
reportlab==3.6.8
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
shutilwhich==1.1.0
six==1.16.0
sklearn==0.0.post11
snuggs==1.4.7
SQLAlchemy==2.0.5.post1
ssh-import-id==5.11
stack-data==0.6.3
systemd-python==234
tabulate==0.9.0
terminator==2.1.1
threadpoolctl==3.2.0
tk==0.1.0
tomli==2.0.1
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.10
webencodings==0.5.1
Werkzeug==2.2.3
xdg==5
xkit==0.0.0
zipp==1.0.0
